# services/identity/openapi/openapi.yaml
openapi: "3.1.0"
info:
  title: AIAUDITOR Identity Service
  version: "0.1.0"
  description: |
    Authentication, JWT RS256 issuance, 7-persona RBAC, user management,
    and organization management for the AIAUDITOR platform.

servers:
  - url: http://localhost:8081
    description: Local development

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: UNAUTHORIZED
        message:
          type: string
          example: invalid or expired token

    TokenResponse:
      type: object
      required: [access_token, refresh_token, token_type, expires_in]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          description: Access token lifetime in seconds
          example: 900

    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        org_id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        display_name:
          type: string
        is_active:
          type: boolean
        last_login_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    Organization:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        tier:
          type: string
          enum: [standard, professional, enterprise]
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    Role:
      type: object
      properties:
        id:
          type: string
          format: uuid
        org_id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
          enum:
            - internal_auditor
            - cae
            - audit_committee
            - auditee_ciso
            - cosourced_provider
            - ptwg_member
            - beta_tester
        is_system:
          type: boolean
        created_at:
          type: string
          format: date-time

security:
  - bearerAuth: []

paths:
  /health:
    get:
      summary: Service health
      security: []
      responses:
        "200":
          description: Service is healthy

  /auth/register:
    post:
      summary: Register a new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [org_id, email, display_name, password]
              properties:
                org_id:
                  type: string
                  format: uuid
                email:
                  type: string
                  format: email
                display_name:
                  type: string
                password:
                  type: string
                  minLength: 8
                persona:
                  type: string
                  enum: [internal_auditor, cae, audit_committee, auditee_ciso, cosourced_provider, ptwg_member, beta_tester]
      responses:
        "201":
          description: User created
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: User already exists

  /auth/login:
    post:
      summary: Authenticate and receive JWT tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [org_id, email, password]
              properties:
                org_id:
                  type: string
                  format: uuid
                email:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Successful login
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "400":
          description: Bad request
        "401":
          description: Invalid credentials

  /auth/refresh:
    post:
      summary: Rotate refresh token and get new access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        "200":
          description: New token pair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          description: Invalid or expired refresh token

  /auth/logout:
    post:
      summary: Revoke refresh token
      responses:
        "200":
          description: Logged out successfully
        "401":
          description: Unauthorized

  /users:
    get:
      summary: List users in the caller's organization
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden

  /users/{id}:
    get:
      summary: Get a user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found

    put:
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [display_name]
              properties:
                display_name:
                  type: string
      responses:
        "200":
          description: User updated
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found

    delete:
      summary: Deactivate a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: User deactivated
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found

  /users/{id}/roles:
    post:
      summary: Assign a role to a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role_slug]
              properties:
                role_slug:
                  type: string
      responses:
        "200":
          description: Role assigned
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: User or role not found

  /organizations:
    post:
      summary: Create a new organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, slug]
              properties:
                name:
                  type: string
                slug:
                  type: string
                tier:
                  type: string
                  enum: [standard, professional, enterprise]
      responses:
        "201":
          description: Organization created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "401":
          description: Unauthorized
        "409":
          description: Slug already exists

    get:
      summary: List all organizations (admin)
      responses:
        "200":
          description: List of organizations
          content:
            application/json:
              schema:
                type: object
                properties:
                  organizations:
                    type: array
                    items:
                      $ref: "#/components/schemas/Organization"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden

  /organizations/{id}:
    get:
      summary: Get an organization by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Organization found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden (not your org)
        "404":
          description: Not found

    put:
      summary: Update an organization
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                tier:
                  type: string
      responses:
        "200":
          description: Organization updated
        "401":
          description: Unauthorized
        "403":
          description: Forbidden

  /organizations/{id}/members:
    post:
      summary: Add a member to an organization
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, display_name, password]
              properties:
                email:
                  type: string
                  format: email
                display_name:
                  type: string
                password:
                  type: string
                  minLength: 8
                persona:
                  type: string
      responses:
        "201":
          description: Member added
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "409":
          description: User already exists

  /roles:
    get:
      summary: List roles in the caller's organization
      responses:
        "200":
          description: List of roles
          content:
            application/json:
              schema:
                type: object
                properties:
                  roles:
                    type: array
                    items:
                      $ref: "#/components/schemas/Role"
        "401":
          description: Unauthorized